+++ src/main/resources/application.properties
@@ -1,5 +1,10 @@
+#spring.datasource.driver-class-name=
+spring.datasource.url=jdbc:mysql:https://www.db4free.net/phpMyAdmin/db_structure.php?db=fiw_kanban_2018
 #spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://db.f4.htw-berlin.de:5432/_s0553666__fiw_kanban
-spring.datasource.username=_s0553666__fiw_kanban_generic
-spring.datasource.password=fiw-kanban
+#spring.datasource.url=jdbc:postgresql:https://studi.f4.htw-berlin.de/postgresql/_s0553666__fiw_kanban_generic/_fiw_kanban
+#(((spring.datasource.url=jdbc:postgresql:db.f4.htw-berlin.de:5432:require&database=_s0553666__fiw_kanban
+#spring.datasource.username=_s0553666__fiw_kanban_generic
+#spring.datasource.password=fiw-kanban
+spring.datasource.username=fiw_kanban_2018
+spring.datasource.password=fiw-kanban-2018
 spring.jpa.generate-ddl=true
